move columnTypeValidation to components/utils
