+++ src/scripts/modules/components/utils/columnTypeValidation.js
